Externalized all gin handler functions to the middleware package
